Some function shuffling and different handling of ports with strings
