+++ src/as2fm/jani_generator/ros_helpers/ros_communication_handler.py
@@ -19,7 +19,7 @@ Generic class for generators of SCXML state machine for specific ROS communicati
 
 from typing import Dict, Iterator, List, Optional, Type
 
-from as2fm.as2fm_common.array_type import get_default_expression_for_type, value_to_string_expr
+from as2fm.as2fm_common.common import get_default_expression_for_type, value_to_string_expr
 from as2fm.jani_generator.jani_entries import JaniModel
 from as2fm.scxml_converter.ascxml_extensions.ros_entries import RosDeclaration
 from as2fm.scxml_converter.data_types.type_utils import get_data_type_from_string
